The length of a disassemnled insn is always positive. Make DisResult::len an UInt.
git-svn-id: svn://svn.valgrind.org/vex/trunk@3054

+++ VEX/priv/guest_amd64_toIR.c
@@ -31763,7 +31763,7 @@ DisResult disInstr_AMD64_WRK (
    }
 
    DIP("\n");
-   dres.len = (Int)toUInt(delta - delta_start);
+   dres.len = toUInt(delta - delta_start);
    return dres;
 }
